energy_data_extractors/requirements.txt
# Data Pipeline Dependencies - Hexa Climate Dataset Repository
# Python 3.13+ compatible packages for automated data extraction and processing

# Core Data Processing
pandas>=2.1.0
pyarrow>=13.0.0
numpy>=1.24.0

# Excel/CSV File Processing
openpyxl>=3.1.0
xlrd>=2.0.1
xlsxwriter>=3.1.0

# AWS S3 Integration
boto3>=1.29.0
botocore>=1.32.0

# Environment Configuration (SECURITY)
python-dotenv>=1.0.0

# Web Scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# Selenium (for dynamic content extraction)
selenium>=4.15.0

# Data Compression and Archives
zipfile36>=0.1.3

# Date and Time Processing
python-dateutil>=2.8.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Jupyter for data analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# System and OS utilities
psutil>=5.9.0
